% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{Emojis}
\format{
A \code{data.table} object with X rows and X columns:
\describe{
  \item{code}{Unicode code point.}
  \item{char}{Unicode characters.}
  \item{name}{Emoji short name.}
  \item{type}{Emoji category.}
  \item{nchar}{Number of Unicode characters.}
}
}
\source{
\url{https://data.world/kgarrett/emojis}
}
\usage{
emojis
}
\description{
List of emojis.
}
\keyword{datasets}
