% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{EmoticonRegex}
\alias{EmoticonRegex}
\title{Emoticon regex}
\usage{
EmoticonRegex(emoticon.re, front, back)
}
\arguments{
\item{emoticon.re}{Regular expressions for each emoticon.}

\item{front}{Regular expression for symbols that can appear in
front of the emoticon.}

\item{back}{Regular expression for symbols that can appear behind
the emoticon.}
}
\value{
A single regular expression for all emoticons.
}
\description{
Builds emoticon regex
}
