% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoticons.re}
\alias{emoticons.re}
\title{Emoticon Regular Expressions}
\format{
A \code{data.table} object with X rows and X columns:
\describe{
  \item{emoticon}{The emoticon.}
  \item{regex}{The regular expression matching the emoticon.}
  \item{front}{The regular expression matching what can be in front of the emoticon.}
  \item{back}{The regular expression matching what can be behind the emoticon.}
}
}
\usage{
emoticons.re
}
\description{
Regular expressions for detecting emoticons.
}
\keyword{datasets}
