% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{EmoticonsAndEmojis}
\alias{EmoticonsAndEmojis}
\title{Find emoticons and emojis}
\usage{
EmoticonsAndEmojis(
  data,
  text.col = "text",
  cols = NULL,
  include.text = TRUE,
  funcs = list(emoticon = FindEmoticons, emoji = FindUnicodeEmojis, slack =
    FindSlackEmojis)
)
}
\arguments{
\item{data}{Data.table of data to find emoticons and emojis in.}

\item{text.col}{Name of the column containing the text.}

\item{cols}{Columns from data to keep. If NULL, then uses the key
of data.}

\item{include.text}{If TRUE, includes the original text in the
table.}

\item{funcs}{List of functions to use for identifying emoticons and
emojis.}
}
\value{
A data.table object with emoticons and emojis extracted
  from the text and subset of columns from data.
}
\description{
Find emoticons and emojis
}
