% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{EscapeRegex}
\alias{EscapeRegex}
\title{Escape regular expressions}
\usage{
EscapeRegex(re)
}
\arguments{
\item{re}{A regular expression to escape.}
}
\value{
The escaped regular expression.
}
\description{
Escape regular expressions
}
