% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{FindEmoticons}
\alias{FindEmoticons}
\title{Find emoticons}
\usage{
FindEmoticons(text, emoticons = emoticons.re, include.text = TRUE)
}
\arguments{
\item{text}{The text to find emoticons in.}

\item{emoticons}{Regular expressions for detecting emoticons.}

\item{include.text}{If TRUE, includes the original text in the
table.}
}
\value{
A data.tabble object with emoticons extracted from the text.
}
\description{
Find emoticons
}
