% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{FindRegex}
\alias{FindRegex}
\title{Find regex}
\usage{
FindRegex(text, regex, include.text = TRUE)
}
\arguments{
\item{text}{Text to find regular expressions in.}

\item{regex}{The regular expressions.}

\item{include.text}{If TRUE, includes the original text in the
table.}
}
\value{
A data.table object with the result of the regex extracted
  and where it starts and ends in the original text.
}
\description{
Builds a data.table object from locations matching a set of regular
expressions.
}
