% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{FindSlackEmojis}
\alias{FindSlackEmojis}
\title{Find Slack emojis}
\usage{
FindSlackEmojis(text, include.text = TRUE)
}
\arguments{
\item{text}{Text to find emojis in.}

\item{include.text}{If TRUE, includes the original text in the
table.}
}
\value{
A data.table object with emojis extracted from the text.
}
\description{
Find Slack emojis
}
