% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{FindUnicodeEmojisSeq}
\alias{FindUnicodeEmojisSeq}
\title{Find Unicode emoji sequences}
\usage{
FindUnicodeEmojisSeq(text, include.text = TRUE)
}
\arguments{
\item{text}{Text to find emoji sequences in.}

\item{include.text}{If TRUE, includes the original text in the
table.}
}
\value{
A data.table object with emoji sequences extracted from the text.
}
\description{
Find Unicode emoji sequences
}
