% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{FirstSentence}
\alias{FirstSentence}
\title{First sentence}
\usage{
FirstSentence(text, nclusters = parallel::detectCores())
}
\arguments{
\item{text}{Text to use.}

\item{nclusters}{Number of clusters to use for running openNLP to
detect sentences.}
}
\value{
First sentence of text.
}
\description{
Returns first sentence of a piece of text.
}
