% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{LastSentence}
\alias{LastSentence}
\title{Last sentence}
\usage{
LastSentence(text, nclusters = parallel::detectCores())
}
\arguments{
\item{text}{Text to use.}

\item{nclusters}{Number of clusters to use for running openNLP to
detect sentences.}
}
\value{
Last sentence of text.
}
\description{
Returns last sentence of a piece of text.
}
