% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{SplitSentences}
\alias{SplitSentences}
\title{Split sentences}
\usage{
SplitSentences(text, nclusters = parallel::detectCores())
}
\arguments{
\item{text}{Text to split into sentences.}

\item{nclusters}{Number of clusters to use for running openNLP}
}
\value{
List of same size as text with each element a vector of sentences.
}
\description{
Split text into individual sentences using openNLP.
}
