% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoticons.R
\name{TextBeforeAfterEmoticons}
\alias{TextBeforeAfterEmoticons}
\title{Text before and after emoticons}
\usage{
TextBeforeAfterEmoticons(emoticons, nclusters = parallel::detectCores())
}
\arguments{
\item{emoticons}{Emoticon data.table object.}

\item{nclusters}{Number of clusters to use for running openNLP to
detect sentences.}
}
\value{
Emoticon data.table object with \code{text.before},
  \code{text.after}, \code{sentence.before} and
  \code{sentence.after} columns added.
}
\description{
Computes text before and after emoticons.
}
